requirements.txt
uvicorn
fastapi
pydantic
llama-cpp-python
